Fixed present attendance count as per muster roll in attendance report

+++ backend/muster-roll/src/main/java/org/egov/web/models/report/AttendanceReportDetail.java
@@ -57,10 +57,10 @@ public class AttendanceReportDetail {
     private String attendanceMarker;
 
     @JsonProperty("presentDaysOriginal")
-    private Integer presentDaysOriginal;
+    private Double presentDaysOriginal;
 
     @JsonProperty("presentDaysModified")
-    private Integer presentDaysModified;
+    private Double presentDaysModified;
 
     @JsonProperty("dailyAttendance")
     private Map<String, String> dailyAttendance;
